+++ models/Story.js
@@ -1,6 +1,5 @@
 const { Model, DataTypes } = require("sequelize");
 const sequelize = require("../config/connection");
-const multer = require("multer");
 
 class Story extends Model {}
 
@@ -49,17 +48,4 @@ Story.init(
   }
 );
 
-// Configure multer for file uploads
-const storage = multer.diskStorage({
-  destination: function (req, file, cb) {
-    cb(null, "public/images"); // Specify the directory where uploaded images will be stored
-  },
-  filename: function (req, file, cb) {
-    // Generate a unique filename for the uploaded file
-    cb(null, Date.now() + "-" + file.originalname);
-  },
-});
-
-const upload = multer({ storage: storage }).array("image", 5);
-
-module.exports = { Story, upload };
+module.exports = { Story };
